- Converted to a Unity Package - Adjusted namespace to reflect style of other Unity packages - Adjusted menu entry to better fit into the window tab - Added script and editor assembly definitions (required by Unity 2018.3 for packages)
