+++ Editor/Unity.Tools.GraphVisualizer.Editor.asmdef
@@ -1,0 +1,14 @@
+{
+    "name": "Unity.Tools.GraphVisualizer.Editor",
+    "references": [
+        "Unity.Tools.GraphVisualizer"
+    ],
+    "optionalUnityReferences": [],
+    "includePlatforms": [],
+    "excludePlatforms": [],
+    "allowUnsafeCode": false,
+    "overrideReferences": false,
+    "precompiledReferences": [],
+    "autoReferenced": true,
+    "defineConstraints": []
+}
