+++ Editor/Unity.Tools.GraphVisualizer.Editor.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 80d5dfa1b3f9ff70ab8b5c894311cfc0
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
