+++ Scripts/Unity.Tools.GraphVisualizer.asmdef
@@ -1,0 +1,3 @@
+﻿{
+	"name": "Unity.Tools.GraphVisualizer"
+}
