+++ Scripts/Unity.Tools.GraphVisualizer.asmdef.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 408b2a2d5172cc9ffba4a5d8015c95c9
+AssemblyDefinitionImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
