+++ package.json
@@ -1,0 +1,12 @@
+{
+    "name": "com.unity.tools.graphvisualizer",
+    "displayName": "Graph Visualizer",
+    "version": "0.0.1-preview.1",
+    "unity": "2017.4",
+    "description": "The PlayableGraph Visualizer is a tool that displays the PlayableGraphs in the scene. It can be used in both Play and Edit mode and will always reflect the current state of the graph. Playable nodes are represented by colored nodes, varying according to their type. Connections color intensity indicates its weight.",
+    "keywords": [
+        "tools",
+        "unity"
+    ],
+    "category": "Unity"
+}
