+++ package.json.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: bdb853893c29eac458aaed87f5255f9e
+TextScriptImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
